% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{conditional}
\alias{conditional}
\title{Conditional Probabilities (or Counts).}
\usage{
conditional(data, wt, tibble = TRUE, stat = "prob")
}
\arguments{
\item{data}{A data frame.}

\item{wt}{An optional column of weights.}

\item{tibble}{Whether or not to return results in a \code{tibble}, or
something simpler. Default \code{TRUE}.}

\item{stat}{Either "prob" to return probabilities, or "count" to return
counts.}
}
\value{
A \code{tibble} or \code{matrix} of conditional probabilities or
counts.
}
\description{
Calculates the conditional probability (or counts) across all pairwise
sets of columns in a data frame. This is the probability of "A" given "B".
}
\examples{
conditional(pets, wt)
conditional(pets[-5])
conditional(pets, wt, tibble = FALSE)
conditional(pets, wt, stat = "count")

}
