% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{intersection}
\alias{intersection}
\alias{joint}
\title{Probability (or Counts) of Intersection.}
\usage{
intersection(data, wt, tibble = TRUE, stat = "prob")

joint(data, wt, tibble = TRUE, stat = "prob")
}
\arguments{
\item{data}{A data frame.}

\item{wt}{An optional column of weights.}

\item{tibble}{Whether or not to return results in a \code{tibble}, or
something simpler. Default \code{TRUE}.}

\item{stat}{Either "prob" to return probabilities, or "count" to return
counts.}
}
\value{
A \code{tibble} or \code{matrix} of probability or counts of
intersection.
}
\description{
Calculates the probability (or counts) of intersection across all pairwise
sets of columns in a data frame. This is the probability of both "A" and "B"
occurring, otherwise known as a joint probability.
}
\examples{
intersection(pets, wt)
intersection(pets[-5])
intersection(pets, wt, tibble = FALSE)
intersection(pets, wt, stat = "count")


joint(pets, wt)
joint(pets[-5])
joint(pets, wt, tibble = FALSE)
joint(pets, wt, stat = "count")

}
