% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_onezero.R
\name{is_onezero}
\alias{is_onezero}
\title{Check if data is binary.}
\usage{
is_onezero(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Does a check to see if the data contains only 1's, 0's or NA's.
}
\examples{

is_onezero(c(1, 0, NA))
is_onezero(c(1, 0, -1))

dplyr::select_if(pets, is_onezero)

}
