% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Index (or Distance).}
\usage{
jaccard(data, wt, tibble = TRUE, distance = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{wt}{An optional column of weights.}

\item{tibble}{Whether or not to return results in a \code{tibble}, or
something simpler. Default \code{TRUE}.}

\item{distance}{Whether or not to return Jaccard distance. Default
\code{FALSE}.}
}
\value{
A \code{tibble} or \code{matrix} of Jaccard indices or distances.
}
\description{
Calculates the Jaccard similarity index (or distance) across all pairwise
sets of columns in a data frame.
}
\examples{
jaccard(pets, wt)
jaccard(pets[-5])
jaccard(pets, wt, tibble = FALSE)
jaccard(pets, wt, distance = TRUE)

}
