% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal}
\alias{marginal}
\title{Marginal Probabilities (or Counts).}
\usage{
marginal(data, wt, tibble = TRUE, stat = "prob")
}
\arguments{
\item{data}{A data frame.}

\item{wt}{An optional column of weights.}

\item{tibble}{Whether or not to return results in a \code{tibble}, or
something simpler. Default \code{TRUE}.}

\item{stat}{Either "prob" to return probabilities, or "count" to return
counts.}
}
\value{
A \code{tibble} or vector of marginal probabilities or counts.
}
\description{
Calculate marginal probabilities or counts across all columns in a data
frame.
}
\examples{
marginal(pets, wt)
marginal(pets[-5])
marginal(pets, wt, tibble = FALSE)
marginal(pets, wt, stat = "count")

}
