% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onezero.R
\name{onezero}
\alias{onezero}
\title{Convert vector to ones and zeros.}
\usage{
onezero(x, one, zero, keep.na = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{one}{Value in \code{x} to receive values of 1.}

\item{zero}{Value in \code{x} to receive values of 0. If left unspecified,
all remaining values assumed to receive value of 0.}

\item{keep.na}{Only needed if \code{zero} is not specified, whether or not
to retain missing values during conversion, default \code{TRUE}.}
}
\value{
A numeric vector of class \code{"onezero"}.
}
\description{
Converts a vector into ones and zeroes, any values which fall outside of what
is set by the arguments gets converted to missing. This vector also receives
a special class of \code{"onezero"} which makes succeeding functions go quicker.
}
\examples{

a <- c("zero", "one", "two")
onezero(a, "one", "zero")

b <- c(1, -99, -99, 1, 1)
onezero(b, 1, -99)

z <- c("a", "b", "c", NA)
onezero(z, one = "b")
onezero(z, one = "b", keep.na = FALSE)

}
