% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pets.R
\docType{data}
\name{pets}
\alias{pets}
\title{Pet ownership.}
\format{Data frame with 10 rows and 5 columns.
\describe{
   \item{dog}{Owned a dog?}
   \item{cat}{Owned a cat?}
   \item{fish}{Owned a fish?}
   \item{bird}{Owned a bird?}
   \item{wt}{Weight given to respondent.}
}}
\usage{
data(pets)
}
\description{
Ten people were asked if they had ever owned a dog, cat, fish, or bird. The
second person was not asked if he had ever owned a cat.
}
\keyword{datasets}
