% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{union}
\alias{union}
\title{Probability (or Counts) of Union.}
\usage{
union(data, wt, tibble = TRUE, stat = "prob")
}
\arguments{
\item{data}{A data frame.}

\item{wt}{An optional column of weights.}

\item{tibble}{Whether or not to return results in a \code{tibble}, or
something simpler. Default \code{TRUE}.}

\item{stat}{Either "prob" to return probabilities, or "count" to return
counts.}
}
\value{
A \code{tibble} or \code{matrix} of probability or counts of
union.
}
\description{
Calculates the probability (or counts) of union across all pairwise
sets of columns in a data frame. This is the probability of either "A" or
"B" occurring.
}
\examples{
union(pets, wt)
union(pets[-5])
union(pets, wt, tibble = FALSE)
union(pets, wt, stat = "count")

}
