% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Weighted Mean.}
\usage{
weighted_mean(x, w)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector of weights.}
}
\description{
Weighted Mean.
}
